seperate DescrptSeA from ModelSeA
